Move struct template `binder` into `collector` from `gc_pointer`
